frontend/backend/requirements.txt
Flask==2.0.2
Flask-Testing==0.8.1
Flask-WTF==0.15.1
PyJWT==2.3.0
pytest==6.2.5
psycopg2==2.9.2
Flask-Cors==3.0.10
requests==2.22.0
